requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
osmnx==1.7.0
geopandas==0.14.1
shapely==2.0.2
trimesh==4.0.5
numpy==1.26.3
rasterio==1.3.9
networkx==3.2.1
mapbox-earcut==1.0.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
aiofiles==23.2.1
pydantic-settings==2.1.0
requests==2.31.0
lxml==5.1.0
scipy==1.11.4
Pillow==10.2.0
pyrosm==0.6.2
pyarrow==16.1.0
pyogrio==0.10.0
triangle==20250106

